Added 2D DCBC reggie for testing Neumann BCs with one or more symmetries, which require 2x or 4x the deposited charge to account for mirror charges.

+++ src/particles/pic/deposition/pic_depo_tools.f90
@@ -125,7 +125,7 @@ ASSOCIATE( SurfNodeSource => SurfNodeSourceMPI )
     ! Add charge contribution
     IF (SurfNodeSymmetryFactor(NonUniqueNodeID).GT.0) THEN
       SurfNodeSource(iDepoSurfNodeID) = SurfNodeSource(iDepoSurfNodeID) + PartDistDepo(iNode)/DistSum*Charge&
-                                                                          *SurfNodeSymmetryFactor(NonUniqueNodeID)
+                                                                          *REAL(SurfNodeSymmetryFactor(NonUniqueNodeID))
     ELSE
       SurfNodeSource(iDepoSurfNodeID) = SurfNodeSource(iDepoSurfNodeID) + PartDistDepo(iNode)/DistSum*Charge
     END IF ! SurfNodeSymmetryFactor(NonUniqueNodeID).GT.0
